Fix multikey issues on sway : move to bonsai
To fix the problems mentionned in the following ticket I worked on new
solution to handle complex key bindings on mobile devices:

https://todo.sr.ht/~mil/sxmo-tickets/524

This solution is structured as client/server state machine in hare lang.
It read a json transition tree and wait for the client to send events or
to change contexts.

Installing bonsai is ofc required :

$ doas apk add bonsai

Dont forget to $ sxmo_migrate.sh sync interactive to setup the json in
your config folder.

I adapted the migration scripts to not break the json. We cannot
versionise it so let's just check if the user one differ from the sxmo
default one.

+++ scripts/core/sxmo_migrate.sh
@@ -135,10 +135,13 @@ defaultconfig() {
 		cp "$defaultfile" "$userfile"
 		chmod "$filemode" "$userfile"
 	elif [ "$MODE" = "reset" ]; then
-		if [ ! -e "$userfile.needs-migration" ]; then
-			mv "$userfile" "$userfile.needs-migration"
-		else
-			sxmo_log "$userfile was already flagged for needing migration; not overwriting the older one"
+		if [ -e "$userfile" ] \
+			&& ! diff "$defaultfile" "$userfile" > /dev/null; then
+			if [ ! -e "$userfile.needs-migration" ]; then
+				mv "$userfile" "$userfile.needs-migration"
+			else
+				sxmo_log "$userfile was already flagged for needing migration; not overwriting the older one"
+			fi
 		fi
 		cp "$defaultfile" "$userfile"
 		chmod "$filemode" "$userfile"
@@ -248,6 +251,7 @@ sway() {
 	defaultconfig "$(xdg_data_path sxmo/appcfg/sway_template)" "$XDG_CONFIG_HOME/sxmo/sway" 644
 	defaultconfig "$(xdg_data_path sxmo/appcfg/foot.ini)" "$XDG_CONFIG_HOME/foot/foot.ini" 644
 	defaultconfig "$(xdg_data_path sxmo/appcfg/mako.conf)" "$XDG_CONFIG_HOME/mako/config" 644
+	MODE=reset defaultconfig "$(xdg_data_path sxmo/appcfg/bonsai_tree.json)" "$XDG_CONFIG_HOME/sxmo/bonsai_tree.json" 644
 }
 
 xorg() {
